Fix R2025a scheduler count type and record owner validation
Original local commit: e700b15f96a05ce6c60e041e97ae4c93ac223262
Verified source tree: 40ebc31c9431954ba08c8779445ccd6c2f0fac38

+++ +csr/+sim/EventScheduler.m
@@ -42,7 +42,9 @@ classdef EventScheduler < handle
         end
 
         function count = get.PendingCount(obj)
-            count = obj.Active.Count;
+            % Expose a double count consistently with the other counters.
+            % containers.Map.Count returns uint64 on R2025a; IDs stay uint64.
+            count = double(obj.Active.Count);
         end
 
         function id = scheduleAt(obj, time, callback)
